+++ src/main/java/br/com/controlefinanceiro/model/Receita.java
@@ -3,6 +3,7 @@ package br.com.controlefinanceiro.model;
 import br.com.controlefinanceiro.enums.CategoriaReceita;
 import jakarta.persistence.*;
 
+import java.math.BigDecimal;
 import java.time.LocalDateTime;
 
 @Entity
@@ -15,13 +16,13 @@ public class Receita {
     @Column(nullable = false)
     private String descricao;
 
-    private double valor;
+    private BigDecimal valor;
 
     private LocalDateTime data;
 
     private CategoriaReceita categoriaReceita;
 
-    public Receita(Long id, String descricao, double valor, LocalDateTime data, CategoriaReceita categoriaReceita) {
+    public Receita(Long id, String descricao, BigDecimal valor, LocalDateTime data, CategoriaReceita categoriaReceita) {
         this.id = id;
         this.descricao = descricao;
         this.valor = valor;
@@ -45,11 +46,11 @@ public class Receita {
         this.descricao = descricao;
     }
 
-    public double getValor() {
+    public BigDecimal getValor() {
         return valor;
     }
 
-    public void setValor(double valor) {
+    public void setValor(BigDecimal valor) {
         this.valor = valor;
     }
